+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+
+## [0.1.1] - 2019-02-18
 ### Added
 - Function to add default command line args to a parser.
 - DomainPrice command.
@@ -19,5 +21,6 @@ and this project adheres to [Semantic Versioning].
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
+[0.1.1]: https://github.com/Smoose-bv/rrpproxypy/releases/0.1.1
 [0.1.0]: https://github.com/Smoose-bv/rrpproxypy/releases/0.1.0
 [Unreleased]: https://github.com/Smoose-bv/rrpproxypy